moved attribute definitions to openpype/lib
