--- openpype/pipeline/lib/__init__.py
@@ -1,30 +1,0 @@
-from .attribute_definitions import (
-    AbtractAttrDef,
-
-    UIDef,
-    UISeparatorDef,
-    UILabelDef,
-
-    UnknownDef,
-    NumberDef,
-    TextDef,
-    EnumDef,
-    BoolDef,
-    FileDef,
-)
-
-
-__all__ = (
-    "AbtractAttrDef",
-
-    "UIDef",
-    "UISeparatorDef",
-    "UILabelDef",
-
-    "UnknownDef",
-    "NumberDef",
-    "TextDef",
-    "EnumDef",
-    "BoolDef",
-    "FileDef",
-)
